+++ R/global.R
@@ -1,1 +1,0 @@
-globalVariables(c("qdap_dict", "ubuntu_dict"))