+++ tests/testthat/test-wordler.R
@@ -1,5 +1,5 @@
 test_that("Repeated guess letters are compared to target without replacement", {
-  game <- make_new_game()
+  game <- new_wordler()
   game$target <- "SOLAR"
   first_letter_in_pos <- have_a_guess("SOULS", game, allowed_words = c("SOULS"))
   # The repeated `S` should not be counted as in the target
@@ -10,16 +10,9 @@ test_that("Repeated guess letters are compared to target without replacement", {
   expect_equal(last_letter_in_pos$assess[[1]][[4]], "not_in_word")
 })
 
-test_that("Custom dictionary used for target words", {
-  # Words unlikely to be in any default dictionary
-  custom_dict <- c("XXXXX", "ZZZZZ")
-  game <- make_new_game(target_words = custom_dict)
-  expect_true(game$target %in% custom_dict)
-})
-
 test_that("Custom dictionary used to validate guess", {
   # Words unlikely to be in any default dictionary
-  game <- make_new_game(target_words = c("XXXXX"))
+  game <- new_wordler(target = "XXXXX")
   game <- have_a_guess("ZZZZZ", game, "ZZZZZ")
   expect_equal(game$guess_count, 1)
   game <- have_a_guess("ZZZZZ", game, "XXXXX")
@@ -27,7 +20,7 @@ test_that("Custom dictionary used to validate guess", {
 })
 
 test_that("Letters known not to be in word are correctly identified", {
-  game <- make_new_game(target_words = "STAMP")
+  game <- new_wordler(target = "STAMP")
   game <- have_a_guess("TROUT", game)
   expect_equal(game$letters_known_not_in_word, c("R", "O", "U"))
   game <- have_a_guess("STAGE", game)